feat: Upgrade to Astro 7
- Bump astro to 7.0.9 with MDX, Svelte, and Vercel integrations
- Migrate content collections to Content Layer API (src/content.config.ts)
- Update fitness page to use render() from astro:content
- Pin Node 22 via .nvmrc for Astro 6+ requirement
- Upgrade Tailwind/Svelte for Vite 8 compatibility
