+++ package.json
@@ -13,12 +13,12 @@
 		"rag:rebuild": "rm -rf rag/.cache rag/index.json rag/keyword-index.json && npm run rag:build"
 	},
 	"dependencies": {
-		"@astrojs/mdx": "^4.3.9",
-		"@astrojs/svelte": "^7.2.3",
-		"@astrojs/vercel": "^9.0.0",
-		"@tailwindcss/vite": "^4.1.10",
+		"@astrojs/mdx": "7.0.3",
+		"@astrojs/svelte": "9.0.1",
+		"@astrojs/vercel": "11.0.3",
+		"@tailwindcss/vite": "4.3.2",
 		"@xenova/transformers": "^2.17.2",
-		"astro": "^5.15.2",
+		"astro": "7.0.9",
 		"dotenv": "^17.2.3",
 		"gray-matter": "^4.0.3",
 		"jsdom": "^27.0.0",
@@ -29,8 +29,8 @@
 		"p-limit": "^5.0.0",
 		"pdf-parse": "^2.3.12",
 		"sharp": "^0.34.3",
-		"svelte": "^5.46.0",
-		"tailwindcss": "^4.1.10"
+		"svelte": "5.56.5",
+		"tailwindcss": "4.3.2"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^2.2.6",
