+++ src/content.config.ts
@@ -1,0 +1,13 @@
+import { defineCollection } from "astro:content";
+import { glob } from "astro/loaders";
+import { z } from "astro/zod";
+
+const pages = defineCollection({
+	loader: glob({ base: "./src/content/pages", pattern: "**/*.{md,mdx}" }),
+	schema: z.object({
+		title: z.string(),
+		description: z.string().optional(),
+	}),
+});
+
+export const collections = { pages };
