--- src/content/config.ts
@@ -1,11 +1,0 @@
-import { defineCollection, z } from "astro:content";
-
-const pages = defineCollection({
-	type: "content",
-	schema: z.object({
-		title: z.string(),
-		description: z.string().optional(),
-	}),
-});
-
-export const collections = { pages };
