+++ src/pages/fitness.astro
@@ -1,5 +1,5 @@
 ---
-import { getEntry } from "astro:content";
+import { getEntry, render } from "astro:content";
 import MediaGallery from "../components/MediaGallery.astro";
 import MarkdownLayout from "../layouts/MarkdownLayout.astro";
 import PageLayout from "../layouts/PageLayout.astro";
@@ -8,10 +8,10 @@ const raceEntry = await getEntry("pages", "race");
 const outdoorsEntry = await getEntry("pages", "outdoors");
 let RaceContent, OutdoorsContent;
 if (raceEntry) {
-	({ Content: RaceContent } = await raceEntry.render());
+	({ Content: RaceContent } = await render(raceEntry));
 }
 if (outdoorsEntry) {
-	({ Content: OutdoorsContent } = await outdoorsEntry.render());
+	({ Content: OutdoorsContent } = await render(outdoorsEntry));
 }
 ---
